+++ tests/testthat/test-layers.R
@@ -36,23 +36,23 @@ test_that("coloumn of hex colours used", {
   df <- widget_capitals[1:5, ]
   df$colour <- colourvalues::colour_values(1:5, palette = "inferno")
   l <- widget_point( df, fill_colour = "colour" )
-  res <- jsonlite::fromJSON( l$data )
+  res <- jsonify::from_json( l$data )
   expect_true( all(res$properties$fill_colour == colourvalues::colour_values(1:5, palette = "inferno")))
 
   l <- widget_point( df, fill_colour = "colour", legend = FALSE )
-  res <- jsonlite::fromJSON( l$data )
+  res <- jsonify::from_json( l$data )
   expect_true( all(res$properties$fill_colour == colourvalues::colour_values(1:5, palette = "inferno")))
 
   l <- widget_point( df, fill_colour = "#909090" )
-  res <- jsonlite::fromJSON( l$data )
+  res <- jsonify::from_json( l$data )
   expect_true( all(res$properties$fill_colour == "#909090"))
 
   l <- widget_point( df, fill_colour = "1" )
-  res <- jsonlite::fromJSON( l$data )
+  res <- jsonify::from_json( l$data )
   expect_true( all( res$properties$fill_colour == colourvalues::color_values(1) ) )
 
   l <- widget_point( df, fill_colour = 1 )
-  res <- jsonlite::fromJSON( l$data )
+  res <- jsonify::from_json( l$data )
   expect_true( all(res$properties$fill_colour == colourvalues::colour_values(1)))